backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
python-multipart>=0.0.20
aiosqlite>=0.20.0
python-dotenv>=1.0.1
reportlab>=4.0.0
openai>=1.0.0
